[core] move getAttribName to its own header.
