+++ src/Core/Geometry/MeshPrimitives.hpp
@@ -1,4 +1,6 @@
 #pragma once
+
+#include <Core/Geometry/StandardAttribNames.hpp>
 #include <Core/Geometry/TriangleMesh.hpp>
 #include <Core/RaCore.hpp>
 #include <Core/Utils/Color.hpp>
@@ -10,34 +12,6 @@ namespace Geometry {
 //
 // Primitive construction
 //
-
-/// List of standard vertex attributes.
-/// Corresponding standard vertex attribute names are obtained with getAttribName
-/// Information which is in the mesh geometry
-enum MeshData : uint {
-    VERTEX_POSITION,   ///< Vertex positions
-    VERTEX_NORMAL,     ///< Vertex normals
-    VERTEX_TANGENT,    ///< Vertex tangent 1
-    VERTEX_BITANGENT,  ///< Vertex tangent 2
-    VERTEX_TEXCOORD,   ///< U,V  texture coords (last coordinate not used)
-    VERTEX_COLOR,      ///< RGBA color.
-    VERTEX_WEIGHTS,    ///< Skinning weights (not used)
-    VERTEX_WEIGHT_IDX, ///< Associated weight bones
-
-    MAX_DATA
-};
-
-///@{
-/// Get the name expected for a given attrib.
-static constexpr char const* getAttribName[MeshData::MAX_DATA] = { "in_position",
-                                                                   "in_normal",
-                                                                   "in_tangent",
-                                                                   "in_bitangent",
-                                                                   "in_texcoord",
-                                                                   "in_color",
-                                                                   "in_weight",
-                                                                   "in_weight_idx" };
-///@}
 
 /// Create a 2D grid mesh with given number of row and columns
 RA_CORE_API TriangleMesh makePlaneGrid( const uint rows         = 1,
